Fix warnings with Perl 5.20
"Possible precedence issue with control flow operator
 at Shell/EnvImporter/Result.pm line 88"

This one is due to the fact that "return ..." binds more strongly than
"and", so the function would simply "return $self->shell_status == 0",
disregarding $self->command_status and $self->env_status.
Changing "and" to "&&" solves this issue.

"Use of uninitialized value $_[1] in read at IO/Handle.pm"

This is because we don't initialize the hash "%buf" into which we read.
Initializing the relevant keys with the empty string solves this issue

Bug: https://bugs.gentoo.org/526188
Bug: https://rt.cpan.org/Public/Bug/Display.html?id=86171
Bug: https://github.com/gentoo-perl/g-cpan/issues/6